requirements.txt
PyYAML~=5.3.1
requests~=2.25.0
